@@ -1,34 +1,47 @@
-[tool.poetry]
-name = "elia-chat"
-version = "0.5.0"
+[project]
+name = "elia"
+version = "1.0.0b1"
 description = "A terminal interface to ChatGPT."
-authors = ["Darren Burns <darrenb900@gmail.com>"]
+authors = [
+    { name = "Darren Burns", email = "darrenb900@gmail.com" }
+]
+dependencies = [
+    "textual[syntax]==0.58.1",
+    "sqlmodel>=0.0.9",
+    "humanize>=4.6.0",
+    "click>=8.1.6",
+    "xdg-base-dirs>=6.0.1",
+    "pydantic-settings>=2.2.1",
+    "aiosqlite>=0.20.0",
+    "click-default-group>=1.2.4",
+    "litellm>=1.35.38",
+]
 readme = "README.md"
+requires-python = ">= 3.11"
 
-[tool.poetry.dependencies]
-python = "^3.10"
-openai = "^0.27.5"
-textual = "0.46.0"
-# textual = { path = "../textual", develop = true}
-sqlmodel = "^0.0.8"
-humanize = "^4.6.0"
-tiktoken = "^0.4.0"
-click = "^8.1.6"
-langchain = "^0.0.331"
-
-[tool.poetry.scripts]
+[project.scripts]
 elia = "elia_chat.__main__:cli"
 
-[tool.poetry.group.dev.dependencies]
-black = "^23.3.0"
-mypy = "^1.3.0"
-types-peewee = "^3.16.0.0"
-pre-commit = "^3.3.2"
-textual-dev = "^1.0.1"
-
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.rye]
+managed = true
+dev-dependencies = [
+    "black>=23.3.0",
+    "mypy>=1.3.0",
+    "types-peewee>=3.16.0.0",
+    "pre-commit>=3.3.2",
+    "textual-dev>=1.0.1",
+    "pyinstrument>=4.6.2",
+]
 
 [tool.mypy]
 ignore_missing_imports = true
+
+[tool.hatch.metadata]
+allow-direct-references = true
+
+[tool.hatch.build.targets.wheel]
+packages = ["elia_chat"]
