@@ -2,32 +2,62 @@
 Elia CLI
 """
 
+import asyncio
 import pathlib
-from typing import Tuple
+from textwrap import dedent
+import tomllib
+from typing import Any, Tuple
 
 import click
+from click_default_group import DefaultGroup
+
+from rich.console import Console
 
 from elia_chat.app import Elia
-from elia_chat.database.create_database import create_database
+from elia_chat.config import LaunchConfig
 from elia_chat.database.import_chatgpt import import_chatgpt_data
-from elia_chat.database.models import sqlite_file_name
-from elia_chat.models import EliaContext
-from elia_chat.widgets.chat_options import DEFAULT_MODEL, MODEL_MAPPING
+from elia_chat.database.database import create_database, sqlite_file_name
+from elia_chat.launch_args import QuickLaunchArgs
+from elia_chat.locations import config_file
+
+console = Console()
 
 
-@click.group(invoke_without_command=True)
-@click.pass_context
-def cli(context: click.Context) -> None:
-    """
-    Elia: A terminal ChatGPT client built with Textual
-    """
-    app = Elia(context=None)
-    # Run the app if no subcommand is provided
-    if context.invoked_subcommand is None:
-        # Create the database if it doesn't exist
-        if sqlite_file_name.exists() is False:
-            create_database()
-        app.run()
+def create_db_if_not_exists() -> None:
+    if not sqlite_file_name.exists():
+        click.echo(f"Creating database at {sqlite_file_name!r}")
+        asyncio.run(create_database())
+
+
+def load_or_create_config_file() -> dict[str, Any]:
+    config = config_file()
+
+    try:
+        file_config = tomllib.loads(config.read_text())
+    except FileNotFoundError:
+        file_config = {}
+        try:
+            config.touch()
+        except OSError:
+            pass
+
+    return file_config
+
+
+@click.group(cls=DefaultGroup, default="default", default_if_no_args=True)
+def cli() -> None:
+    """Interact with large language models using your terminal."""
+
+
+@cli.command()
+@click.argument("prompt", nargs=-1, type=str, required=False)
+def default(prompt: tuple[str, ...]):
+    prompt = prompt or ("",)
+    joined_prompt = " ".join(prompt)
+    create_db_if_not_exists()
+    file_config = load_or_create_config_file()
+    app = Elia(LaunchConfig(**file_config), startup_prompt=joined_prompt)
+    app.run()
 
 
 @cli.command()
@@ -38,9 +68,28 @@ def reset() -> None:
     This command will delete the database file and recreate it.
     Previously saved conversations and data will be lost.
     """
-    sqlite_file_name.unlink(missing_ok=True)
-    create_database()
-    click.echo(f"♻️  Database reset @ {sqlite_file_name}")
+    from rich.padding import Padding
+    from rich.text import Text
+
+    console.print(
+        Padding(
+            Text.from_markup(
+                dedent(f"""\
+[u b red]Warning![/]
+
+[b red]This will delete all messages and chats.[/]
+
+You may wish to create a backup of \
+"[bold blue u]{str(sqlite_file_name.resolve().absolute())}[/]" before continuing.
+            """)
+            ),
+            pad=(1, 2),
+        )
+    )
+    if click.confirm("Delete all chats?", abort=True):
+        sqlite_file_name.unlink(missing_ok=True)
+        asyncio.run(create_database())
+        console.print(f"♻️  Database reset @ {sqlite_file_name}")
 
 
 @cli.command("import")
@@ -57,8 +106,8 @@ def import_file_to_db(file: pathlib.Path) -> None:
     This command will import the ChatGPT conversations from a local
     JSON file into the database.
     """
-    import_chatgpt_data(file=file)
-    click.echo(f"✅  ChatGPT data imported into database {file}")
+    asyncio.run(import_chatgpt_data(file=file))
+    console.print(f"[green]ChatGPT data imported from {str(file)!r}")
 
 
 @cli.command()
@@ -66,16 +115,22 @@ def import_file_to_db(file: pathlib.Path) -> None:
 @click.option(
     "-m",
     "--model",
-    type=click.Choice(choices=list(MODEL_MAPPING.keys())),
-    default=DEFAULT_MODEL.name,
+    type=str,
+    default="gpt-3.5-turbo",
     help="The model to use for the chat",
 )
 def chat(message: Tuple[str, ...], model: str) -> None:
     """
     Start Elia with a chat message
     """
-    context = EliaContext(chat_message=" ".join(message), model_name=model)
-    app = Elia(context=context)
+    quick_launch_args = QuickLaunchArgs(
+        launch_prompt=" ".join(message),
+        launch_prompt_model_name=model,
+    )
+    launch_config = LaunchConfig(
+        default_model=quick_launch_args.launch_prompt_model_name,
+    )
+    app = Elia(launch_config, quick_launch_args.launch_prompt)
     app.run()
 
 
