@@ -1,9 +1,8 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
+import datetime
 
-from langchain.schema import BaseMessage
-from sqlmodel import Session
 from textual import log
 
 from elia_chat.database.converters import (
@@ -11,27 +10,30 @@ from elia_chat.database.converters import (
     chat_message_to_message_dao,
     message_dao_to_chat_message,
 )
-from elia_chat.database.models import ChatDao, MessageDao, engine
-from elia_chat.models import ChatData
+from elia_chat.database.database import get_session
+from elia_chat.database.models import ChatDao, MessageDao
+from elia_chat.models import ChatData, ChatMessage
 
 
 @dataclass
 class ChatsManager:
     @staticmethod
-    def all_chats() -> list[ChatData]:
-        chat_daos = ChatDao.all()
+    async def all_chats() -> list[ChatData]:
+        chat_daos = await ChatDao.all()
         return [chat_dao_to_chat_data(chat) for chat in chat_daos]
 
     @staticmethod
-    def get_chat(chat_id: str) -> ChatData:
-        chat_dao = ChatDao.from_id(chat_id)
+    async def get_chat(chat_id: int) -> ChatData:
+        chat_dao = await ChatDao.from_id(chat_id)
         return chat_dao_to_chat_data(chat_dao)
 
     @staticmethod
-    def get_messages(chat_id: str | int) -> list[BaseMessage]:
-        with Session(engine) as session:
+    async def get_messages(
+        chat_id: int,
+    ) -> list[ChatMessage]:
+        async with get_session() as session:
             try:
-                chat: ChatDao | None = session.get(ChatDao, int(chat_id))
+                chat: ChatDao | None = await session.get(ChatDao, chat_id)
             except ValueError:
                 raise RuntimeError(
                     f"Malformed chat ID {chat_id!r}. "
@@ -41,40 +43,56 @@ class ChatsManager:
             if not chat:
                 raise RuntimeError(f"Chat with ID {chat_id} not found.")
             message_daos = chat.messages
-            session.commit()
-        # Convert MessageDao objects to BaseMessages
-        chat_messages = []
+            await session.commit()
+
+        # Convert MessageDao objects to ChatMessages
+        model = chat.model
+        chat_messages: list[ChatMessage] = []
         for message_dao in message_daos:
-            chat_message = message_dao_to_chat_message(message_dao)
+            chat_message = message_dao_to_chat_message(message_dao, model)
             chat_messages.append(chat_message)
 
         log.debug(f"Retrieved {len(chat_messages)} chats for chat {chat_id!r}")
         return chat_messages
 
     @staticmethod
-    def create_chat(chat_data: ChatData) -> int:
+    async def create_chat(chat_data: ChatData) -> int:
         log.debug(f"Creating chat in database: {chat_data!r}")
 
-        chat = ChatDao(model=chat_data.model_name, title="Untitled chat")
-
-        for message in chat_data.messages:
-            new_message = MessageDao(role=message.type, content=message.content)
-            chat.messages.append(new_message)
-
-        with Session(engine) as session:
+        model = chat_data.model_name
+        async with get_session() as session:
+            chat = ChatDao(
+                model=model,
+                title="",
+                started_at=datetime.datetime.now(datetime.UTC),
+            )
             session.add(chat)
-            session.commit()
-            session.refresh(chat)
+            await session.commit()
+
+            chat_id = chat.id
+            for message in chat_data.messages:
+                litellm_message = message.message
+                content = litellm_message["content"]
+                new_message = MessageDao(
+                    chat_id=chat_id,
+                    role=litellm_message["role"],
+                    content=content if isinstance(content, str) else "",
+                    model=model,
+                    timestamp=message.timestamp,
+                )
+                (await chat.awaitable_attrs.messages).append(new_message)
+
+            await session.commit()
 
         return chat.id
 
     @staticmethod
-    def add_message_to_chat(chat_id: str, message: BaseMessage) -> None:
-        with Session(engine) as session:
-            chat: ChatDao | None = session.get(ChatDao, chat_id)
+    async def add_message_to_chat(chat_id: int, message: ChatMessage) -> None:
+        async with get_session() as session:
+            chat: ChatDao | None = await session.get(ChatDao, chat_id)
             if not chat:
                 raise Exception(f"Chat with ID {chat_id} not found.")
-            message_dao = chat_message_to_message_dao(message)
-            chat.messages.append(message_dao)
+            message_dao = chat_message_to_message_dao(message, chat_id)
+            (await chat.awaitable_attrs.messages).append(message_dao)
             session.add(chat)
-            session.commit()
+            await session.commit()
