@@ -2,58 +2,84 @@ import json
 from datetime import datetime
 from pathlib import Path
 
-from sqlalchemy.orm import Session
+from rich.console import Console
+from rich.live import Live
+from rich.text import Text
 
-from elia_chat.database.models import MessageDao, ChatDao, engine
+from elia_chat.database.database import get_session
+from elia_chat.database.models import MessageDao, ChatDao
 
 
-def import_chatgpt_data(file: Path) -> None:
-    # Load the JSON data
+async def import_chatgpt_data(file: Path) -> None:
+    console = Console()
+
     with open(file, "r") as f:
         data = json.load(f)
 
-    # Create a new session
-    session = Session(engine)
+    console.print("[green]Loaded and parsed JSON.")
 
-    # Iterate over each chat in the data
-    for chat_data in data:
-        # Create a new ChatDao instance and add it to the session
-        chat = ChatDao(
-            model=None, started_at=datetime.fromtimestamp(chat_data["create_time"])
+    def output_progress(
+        imported_count: int, total_count: int, message_count: int
+    ) -> Text:
+        style = "green" if imported_count == total_count else "yellow"
+        return Text.from_markup(
+            f"Imported [b]{imported_count}[/] of [b]{total_count}[/] chats.\n"
+            f"[b]{message_count}[/] messages in total.",
+            style=style,
         )
-        session.add(chat)
-        session.commit()  # Make sure to commit so that chat.id is assigned
 
-        for message_id, message_data in chat_data["mapping"].items():
-            message_info = message_data.get("message")
-            if message_info:
-                metadata = message_info.get("metadata", {})
-                if metadata:
-                    model = metadata.get("model_slug")
-                    if model is not None:
-                        chat.model = model
-                        session.add(chat)
-                        session.commit()
-
-                message = MessageDao(
-                    chat_id=chat.id,
-                    role=message_info["author"]["role"],
-                    content=message_info["content"]["parts"][0],
-                    timestamp=datetime.fromtimestamp(message_info["create_time"]),
-                    status=message_info.get("status"),
-                    end_turn=message_info.get("end_turn"),
-                    weight=message_info.get("weight"),
-                    meta=metadata,
-                    recipient=message_info.get("recipient"),
+    message_count = 0
+    with Live(output_progress(0, len(data), message_count)) as live:
+        async with get_session() as session:
+            for chat_number, chat_data in enumerate(data, start=1):
+                chat = ChatDao(
+                    title=chat_data.get("title"),
+                    model="gpt-3.5-turbo",
+                    started_at=datetime.fromtimestamp(
+                        chat_data.get("create_time", 0) or 0
+                    ),
                 )
-                session.add(message)
+                session.add(chat)
+                await (
+                    session.commit()
+                )  # Make sure to commit so that chat.id is assigned
 
-        session.commit()
+                for _message_id, message_data in chat_data["mapping"].items():
+                    message_info = message_data.get("message")
+                    if message_info:
+                        metadata = message_info.get("metadata", {})
+                        model = "gpt-3.5-turbo"
+                        if metadata:
+                            model = metadata.get("model_slug")
+                            chat.model = (
+                                "gpt-4-turbo" if model == "gpt-4" else "gpt-3.5-turbo"
+                            )
+                            session.add(chat)
+                            await session.commit()
 
-    # Close the session when you're done
-    session.close()
+                        role = message_info["author"]["role"]
+                        chat_id = chat.id
+                        message = MessageDao(
+                            chat_id=chat_id,
+                            role=role,
+                            content=str(message_info["content"].get("parts", [""])[0]),
+                            timestamp=datetime.fromtimestamp(
+                                message_info.get("create_time", 0) or 0
+                            ),
+                            model=model,
+                            meta=metadata,
+                        )
+                        session.add(message)
+                        message_count += 1
+                        live.update(
+                            output_progress(chat_number, len(data), message_count)
+                        )
+
+                await session.commit()
 
 
 if __name__ == "__main__":
     path = Path("resources/conversations.json")
-    import_chatgpt_data(path)
+    import asyncio
+
+    asyncio.run(import_chatgpt_data(path))
