@@ -1,46 +1,68 @@
-from __future__ import annotations
-
-import pathlib
 from datetime import datetime
-from typing import Any
+from typing import Any, Optional
 
 from sqlalchemy import Column, DateTime, func, JSON, desc
+from sqlalchemy.ext.asyncio import AsyncAttrs
 from sqlalchemy.orm import selectinload
-from sqlmodel import SQLModel, Field, create_engine, Session, select, Relationship
+from sqlmodel import Field, Relationship, SQLModel, select
+
+from elia_chat.database.database import get_session
 
 
-class MessageDao(SQLModel, table=True):
+class SystemPromptsDao(AsyncAttrs, SQLModel, table=True):
+    __tablename__ = "system_prompt"
+
+    id: int | None = Field(default=None, primary_key=True)
+    title: str
+    prompt: str
+    created_at: datetime | None = Field(
+        sa_column=Column(DateTime(), server_default=func.now())
+    )
+
+
+class MessageDao(AsyncAttrs, SQLModel, table=True):
     __tablename__ = "message"
 
-    id: int = Field(default=None, primary_key=True)
-    chat_id: int | None = Field(foreign_key="chat.id")
-    chat: ChatDao = Relationship(back_populates="messages")
+    id: int | None = Field(default=None, primary_key=True)
+    chat_id: Optional[int] = Field(foreign_key="chat.id")
+    chat: Optional["ChatDao"] = Relationship(back_populates="messages")
     role: str
     content: str
     timestamp: datetime | None = Field(
-        sa_column=Column(DateTime(timezone=True), server_default=func.now())
+        sa_column=Column(DateTime(), server_default=func.now())
     )
-    status: str | None
-    end_turn: bool | None
-    weight: float | None
-    meta: dict = Field(sa_column=Column(JSON), default={})
-    recipient: str | None
+    meta: dict[Any, Any] = Field(sa_column=Column(JSON), default={})
+    parent_id: Optional[int] = Field(
+        foreign_key="message.id", default=None, nullable=True
+    )
+    parent: Optional["MessageDao"] = Relationship(
+        back_populates="replies",
+        sa_relationship_kwargs={"remote_side": "MessageDao.id"},
+    )
+    """The message this message is responding to."""
+    replies: list["MessageDao"] = Relationship(back_populates="parent")
+    """The replies to this message
+    (could be multiple replies e.g. from different models).
+    """
+    model: str | None
+    """The model that wrote this response. (Could switch models mid-chat, possibly)"""
 
 
-class ChatDao(SQLModel, table=True):
+class ChatDao(AsyncAttrs, SQLModel, table=True):
     __tablename__ = "chat"
 
     id: int = Field(default=None, primary_key=True)
     model: str
     title: str | None
     started_at: datetime | None = Field(
-        sa_column=Column(DateTime(timezone=True), server_default=func.now())
+        sa_column=Column(DateTime(), server_default=func.now())
     )
     messages: list[MessageDao] = Relationship(back_populates="chat")
+    archived: bool = Field(default=False)
 
     @staticmethod
-    def all() -> list[ChatDao]:
-        with Session(engine) as session:
+    async def all() -> list["ChatDao"]:
+        async with get_session() as session:
             # Create a subquery that finds the maximum
             # (most recent) timestamp for each chat.
             max_timestamp: Any = func.max(MessageDao.timestamp).label("max_timestamp")
@@ -56,22 +78,16 @@ class ChatDao(SQLModel, table=True):
                 .order_by(desc(subquery.c.max_timestamp))
                 .options(selectinload(ChatDao.messages))
             )
-            results = session.exec(statement)
+            results = await session.exec(statement)
             return list(results)
 
     @staticmethod
-    def from_id(chat_id: str) -> ChatDao:
-        with Session(engine) as session:
+    async def from_id(chat_id: int) -> "ChatDao":
+        async with get_session() as session:
             statement = (
                 select(ChatDao)
                 .where(ChatDao.id == int(chat_id))
                 .options(selectinload(ChatDao.messages))
             )
-            result = session.exec(statement).one()
-            return result
-
-
-_this_dir = pathlib.Path(__file__).resolve().parent
-sqlite_file_name = _this_dir / "elia.sqlite"
-sqlite_url = f"sqlite:///{sqlite_file_name}"
-engine = create_engine(sqlite_url)
+            result = await session.exec(statement)
+            return result.one()
